Added basic pipes and flanterm driver.

+++ include/mb_tags.h
@@ -101,7 +101,7 @@ typedef struct {
     uint32_t framebuffer_width;
     uint32_t framebuffer_height;
     uint8_t framebuffer_bpp;
-    uint8_t framebuffer_type;
+    uint16_t framebuffer_type;
     uint8_t reserved;
 
     uint8_t color_info[6];
